langgraph_course/requirements.txt
langchain_community
langchain_openai
python_dotenv